\name{RichPoreTK-package}
\alias{RichPoreTK-package}
\alias{RichPoreTK}
\docType{package}
\title{Tools for processing fast5 files from Oxford Nanopore Devices}
\description{R functions and a command-line interface for processing FAST5 files}
\details{
  \tabular{ll}{
    Package: \tab RichPoreTK\cr
    Type: \tab Package\cr
    Version: \tab 1.0.5\cr
    Date: \tab 2015-05-25\cr
    License: \tab Artistic-2.0\cr
  }
  ~~ An overview of how to use the package, including the most important ~~
  ~~ functions ~~
}
\author{David Eccles (gringer) <bioinformatics@gringene.org>}
\references{
  ~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
  ~~ Optional links to other man pages, e.g. ~~
  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
